fix tidy complaint

+++ comments_test.cpp
@@ -1,3 +1,3 @@
-double return_42() { return 42; }
+int return_42() { return 42; }
 
 int main() { return return_42(); }
